ref: Introduce BaseBackend
