+++ packages/core/src/index.ts
@@ -2,7 +2,8 @@ export { logger } from './logger';
 export { captureException, captureMessage, configureScope } from '@sentry/minimal';
 export { Hub, Scope } from '@sentry/hub';
 export { API } from './api';
-export { BackendClass, BaseClient } from './base';
+export { BaseClient } from './baseclient';
+export { BackendClass, BaseBackend } from './basebackend';
 export { Dsn } from './dsn';
 export { SentryError } from './error';
 export { RequestBuffer } from './requestbuffer';
