+++ packages/core/test/mocks/client.ts
@@ -1,4 +1,4 @@
-import { BaseClient } from '../../src/base';
+import { BaseClient } from '../../src/baseclient';
 import { initAndBind } from '../../src/sdk';
 import { TestBackend, TestOptions } from './backend';
