+++ packages/node/src/backend.ts
@@ -1,6 +1,6 @@
-import { Backend, Dsn, Options, RequestBuffer, SentryError } from '@sentry/core';
+import { BaseBackend, Dsn, Options, SentryError } from '@sentry/core';
 import { getCurrentHub } from '@sentry/hub';
-import { SentryEvent, SentryEventHint, SentryResponse, Severity, Transport } from '@sentry/types';
+import { SentryEvent, SentryEventHint, SentryResponse, Severity } from '@sentry/types';
 import { isError, isPlainObject } from '@sentry/utils/is';
 import { limitObjectDepthToSize, serializeKeysToEventMessage } from '@sentry/utils/object';
 import * as md5 from 'md5';
@@ -23,16 +23,7 @@ export interface NodeOptions extends Options {
 }
 
 /** The Sentry Node SDK Backend. */
-export class NodeBackend implements Backend {
-  /** Creates a new Node backend instance. */
-  public constructor(private readonly options: NodeOptions = {}) {}
-
-  /** Cached transport used internally. */
-  private transport?: Transport;
-
-  /** A simple buffer holding all requests. */
-  private readonly buffer: RequestBuffer<SentryResponse> = new RequestBuffer();
-
+export class NodeBackend extends BaseBackend<NodeOptions> {
   /**
    * @inheritDoc
    */
@@ -97,7 +88,7 @@ export class NodeBackend implements Backend {
     let dsn: Dsn;
 
     if (!this.options.dsn) {
-      throw new SentryError('Cannot sendEvent without a valid Dsn');
+      throw new SentryError('Cannot sendEvent without a valid DSN');
     } else {
       dsn = new Dsn(this.options.dsn);
     }
@@ -112,23 +103,5 @@ export class NodeBackend implements Backend {
     }
 
     return this.transport.captureEvent(event);
-  }
-
-  /**
-   * @inheritDoc
-   */
-  public storeBreadcrumb(): boolean {
-    return true;
-  }
-
-  /**
-   * @inheritDoc
-   */
-  public storeScope(): void {
-    // Noop
-  }
-
-  public getBuffer(): RequestBuffer<SentryResponse> {
-    return this.buffer;
   }
 }
